faenas/requirements.txt
Pillow==6.2.1
django==3.0.1
django-ckeditor==5.8.0
